[vim] Implement b:nolasteditpos for gitcommit ft
This allows a filetype to declare it wants to escape the "transport to
the last editing position" mechanism. For example, when committing with
git, the file always has the name '.git/COMMIT_EDITMSG'. When we exit
the buffer, the cursor is never at the first column of the first line,
but this position is remembered for next time. This is annoying as we
always want to have our cursor at the very beginning, ready to enter
the commit message.